+++ crates/storage/src/storage_drop.rs
@@ -1,7 +1,7 @@
-use crate::storage::Storage;
 use crate::storage_data::StorageData;
+use crate::storage_impl::StorageImpl;
 
-impl<T: StorageData> Drop for Storage<T> {
+impl<T: StorageData> Drop for StorageImpl<T> {
     fn drop(&mut self) {
         if self.apply_wal().is_ok() {
             let _ = self.data.clear_wal();
