+++ src/graph/graph_data_storage.rs
@@ -2,15 +2,14 @@ use super::graph_data::GraphData;
 use super::graph_data_storage_indexes::GraphDataStorageIndexes;
 use crate::storage::StorageVec;
 use crate::DbError;
-use storage::FileStorageData;
+use storage::FileStorage;
 use storage::Storage;
-use storage::StorageData;
 
-pub(crate) struct GraphDataStorage<Data = FileStorageData>
+pub(crate) struct GraphDataStorage<Data = FileStorage>
 where
-    Data: StorageData,
+    Data: Storage,
 {
-    pub(super) storage: std::rc::Rc<std::cell::RefCell<Storage<Data>>>,
+    pub(super) storage: std::rc::Rc<std::cell::RefCell<Data>>,
     pub(super) index: i64,
     #[allow(dead_code)]
     pub(super) indexes: GraphDataStorageIndexes,
@@ -22,7 +21,7 @@ where
 
 impl<Data> GraphData for GraphDataStorage<Data>
 where
-    Data: StorageData,
+    Data: Storage,
 {
     fn capacity(&self) -> Result<u64, crate::DbError> {
         Ok(self.from.len())
